+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 * ***arrays***: new interface `indexes_sym2lt` to convert the coordinates (2-4) in a nD symmetric tensor to the linear storage index of the lower-triangular block.
+* ***vibronic***: new `extrapolate_geom` to compute the extrapolated minimum geometry for vertical vibronic models (VG, VH).
 
 ### Fixed
 * ***vibronic***: Extrapolated shift vector (VH, VG) now supports imaginary frequencies given as negative numbers.